+++ helper/tests/helper-tests-allavro/src/test/resources/allavro/OuterRecordWithInferredInnerRecordNamespace.avsc
