+++ parser/src/test/resources/schemas/TestRecordWithInternalNonNullNamespaceReference.avsc
