+++ parser/src/test/resources/schemas/TestRecordWithInternalNullNamespaceReference.avsc
